+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.10.6] - 2025-10-27
+
+### Fixed
+
+- Add support for width of ButtonComponent
+
 ## [0.10.5] - 2025-10-26
 
 ### Added